+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Add report page tooltip support
 * Add font settings for axis
 * Extended "Chord" settings
+* Add axis background settings
 
 ### Code improvements
 * Fix rendering when "Clear selection" is clicked